// src/app/models/facture.model.ts
export class Facture {
  id?: number;//
  card_number: string;//
  exp_year: string;//
  exp_month: string;//
  cvc: string;
  name: string;
  last_name: string;
  email?: string;
  phone: string;
  doc_number: string;//
  city: string;//
  address: string;
  cellPhone: string;//
  bill: string;//
  value: string;
  expense_id?: number;
  fee_id?: number;
}